requirements.txt
streamlit
pandas
PyYAML

